# dockstore-webservice/src/main/resources/openapi3/openapi.yaml
components:
  schemas:
...
    FileFormat:
      properties:
        value:
          type: string
      type: object
...
    Profile:
      properties:
        avatarURL:
          type: string
        bio:
          type: string
        company:
          type: string
        email:
          type: string
        location:
          type: string
        name:
          type: string
        username:
          type: string
      type: object
...
    SourceFile:
      properties:
        absolutePath:
          type: string
        content:
          type: string
        frozen:
          type: boolean
        id:
          format: int64
          type: integer
        path:
          type: string
        type:
          enum:
          - DOCKSTORE_CWL
          - DOCKSTORE_WDL
          - DOCKERFILE
          - CWL_TEST_JSON
          - WDL_TEST_JSON
          - NEXTFLOW
          - NEXTFLOW_CONFIG
          - NEXTFLOW_TEST_PARAMS
          - DOCKSTORE_YML
          - DOCKSTORE_SERVICE_YML
          - DOCKSTORE_SERVICE_TEST_JSON
          - DOCKSTORE_SERVICE_OTHER
          type: string
        verifiedBySource:
          additionalProperties:
            $ref: '#/components/schemas/VerificationInformation'
          type: object
      type: object
...
    Tag:
      properties:
        automated:
          type: boolean
        commitID:
          type: string
        cwl_path:
          type: string
        dbUpdateDate:
          format: date-time
          type: string
        dirtyBit:
          type: boolean
        dockerfile_path:
          type: string
        doiStatus:
          enum:
          - NOT_REQUESTED
          - REQUESTED
          - CREATED
          type: string
        doiURL:
          type: string
        frozen:
          type: boolean
        hidden:
          type: boolean
        id:
          format: int64
          type: integer
        image_id:
          type: string
        input_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        last_built:
          format: date-time
          type: string
        name:
          type: string
        output_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        reference:
          type: string
        referenceType:
          enum:
          - COMMIT
          - TAG
          - BRANCH
          - NOT_APPLICABLE
          - UNSET
          type: string
        size:
          format: int64
          type: integer
        sourceFiles:
          items:
            $ref: '#/components/schemas/SourceFile'
          type: array
          uniqueItems: true
        valid:
          type: boolean
        validations:
          items:
            $ref: '#/components/schemas/Validation'
          type: array
          uniqueItems: true
        verified:
          type: boolean
        verifiedSource:
          type: string
        verifiedSources:
          items:
            type: string
          type: array
        versionEditor:
          $ref: '#/components/schemas/User'
        wdl_path:
          type: string
        workingDirectory:
          type: string
      type: object
...
    User:
      properties:
        avatarUrl:
          type: string
        curator:
          type: boolean
        id:
          format: int64
          type: integer
        isAdmin:
          type: boolean
        name:
          type: string
        privacyPolicyVersion:
          enum:
          - NONE
          - PRIVACY_POLICY_VERSION_2_5
          type: string
        privacyPolicyVersionAcceptanceDate:
          format: date-time
          type: string
        setupComplete:
          type: boolean
        tosacceptanceDate:
          format: date-time
          type: string
        tosversion:
          enum:
          - NONE
          - TOS_VERSION_1
          type: string
        tosversionAcceptanceDate:
          format: date-time
          type: string
          writeOnly: true
        userProfiles:
          additionalProperties:
            $ref: '#/components/schemas/Profile'
          type: object
        username:
          type: string
      type: object
    Validation:
      properties:
        id:
          format: int64
          type: integer
        message:
          type: string
        type:
          enum:
          - DOCKSTORE_CWL
          - DOCKSTORE_WDL
          - DOCKERFILE
          - CWL_TEST_JSON
          - WDL_TEST_JSON
          - NEXTFLOW
          - NEXTFLOW_CONFIG
          - NEXTFLOW_TEST_PARAMS
          - DOCKSTORE_YML
          - DOCKSTORE_SERVICE_YML
          - DOCKSTORE_SERVICE_TEST_JSON
          - DOCKSTORE_SERVICE_OTHER
          type: string
        valid:
          type: boolean
      type: object
    VerificationInformation:
      properties:
        metadata:
          type: string
        platformVersion:
          type: string
        verified:
          type: boolean
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  contact:
    email: theglobalalliance@genomicsandhealth.org
    name: Dockstore@ga4gh
    url: https://discuss.dockstore.org/t/opening-helpdesk-tickets/1506
  description: The Dockstore API, includes proprietary and GA4GH V1+V2 endpoints
  license:
    name: Apache 2.0
    url: https://github.com/dockstore/dockstore/blob/develop/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Dockstore
  version: 1.6.0
openapi: 3.0.1
paths:
...
  /containers/path/{containerId}/tags:
    get:
      operationId: getTagsByPath
      parameters:
      - in: path
        name: containerId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Tag'
                type: array
                uniqueItems: true
          description: default response
...
  /containers/{containerId}/requestDOI/{tagId}:
    post:
      operationId: requestDOIForToolTag
      parameters:
      - in: path
        name: containerId
        required: true
        schema:
          format: int64
          type: integer
      - in: path
        name: tagId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Tag'
                type: array
                uniqueItems: true
          description: default response
...
  /containers/{containerId}/tags:
    post:
      operationId: addTags
      parameters:
      - in: path
        name: containerId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Tag'
                type: array
                uniqueItems: true
          description: default response
    put:
      operationId: updateTags
      parameters:
      - in: path
        name: containerId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Tag'
                type: array
                uniqueItems: true
          description: default response
  /containers/{containerId}/tags/{tagId}:
    delete:
      operationId: deleteTags
      parameters:
      - in: path
        name: containerId
        required: true
        schema:
          format: int64
          type: integer
      - in: path
        name: tagId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json: {}
          description: default response
...
  /containers/{containerId}/verify/{tagId}:
    post:
      operationId: verifyToolTag
      parameters:
      - in: path
        name: containerId
        required: true
        schema:
          format: int64
          type: integer
      - in: path
        name: tagId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Tag'
                type: array
                uniqueItems: true
          description: default response
...